Fix issue with loading emotes containing a colon
