+++ DiscordBot/appsettings.Development.json
@@ -25,15 +25,28 @@
     "AutoPostChannelID": 997905078121943130
   },
   "RandomReactions": {
-    "FollowupEmotes": {
-      "\uD83D\uDE14": 1
-    },
-    "RandomEmotes": {
-      "\uD83D\uDE14": 0.01
-    },
-    "WelcomeEmotes": {
-      "\uD83D\uDC4B": 0.3
-    },
+    "FollowupEmotes": [
+      {
+        "Emote": "\uD83D\uDE14",
+        "Chance": 1
+      },
+      {
+        "Emote": "<:kaileena:1032782594246451320>",
+        "Chance": 1
+      }
+    ],
+    "RandomEmotes": [
+      {
+        "Emote": "\uD83D\uDE14",
+        "Chance": 1
+      }
+    ],
+    "WelcomeEmotes": [
+      {
+        "Emote": "\uD83D\uDE14",
+        "Chance": 1
+      }
+    ],
     "WelcomeTriggers": [
       "welcome",
       "hello",
