+++ DiscordBot/appsettings.Production.json
@@ -31,28 +31,76 @@
   },
 
   "RandomReactions": {
-    "FollowupEmotes": {
-      "<:pepeComfy:966290245668700191>": 0.2,
-      "\uD83D\uDE14": 0.1,
-      "<:FeelsCoffeeMan:557618333990780959>": 0.2,
-      "<:hype:527072821135015936>": 0.05,
-      "<:iAmAgony:880963270318649344>": 0.05,
-      "<:despair:881299750631145492>": 0.1,
-      "<:KEKyu:852593520798990376>": 0.2,
-      "<:kekPoint:963854834795184158>": 0.2,
-      "<a:partyParrot:792709626726318091>": 0.8
-    },
-    "RandomEmotes": {
-      "<:pepeComfy:966290245668700191>": 0.005,
-      "\uD83D\uDE14": 0.001,
-      "<:FeelsCoffeeMan:557618333990780959>": 0.005,
-      "<:hype:527072821135015936>": 0.001,
-      "<:iAmAgony:880963270318649344>": 0.001,
-      "<:despair:881299750631145492>": 0.002
-    },
-    "WelcomeEmotes": {
-      "<:wave:565223468962938902>": 0.3
-    },
+    "FollowupEmotes": [
+      {
+        "Emote": "<:pepeComfy:966290245668700191>",
+        "Chance": 0.2
+      },
+      {
+        "Emote": "\uD83D\uDE14",
+        "Chance": 0.1
+      },
+      {
+        "Emote": "<:FeelsCoffeeMan:557618333990780959>",
+        "Chance": 0.2
+      },
+      {
+        "Emote": "<:hype:527072821135015936>",
+        "Chance": 0.05
+      },
+      {
+        "Emote": "<:iAmAgony:880963270318649344>",
+        "Chance": 0.05
+      },
+      {
+        "Emote": "<:despair:881299750631145492>",
+        "Chance": 0.1
+      },
+      {
+        "Emote": "<:KEKyu:852593520798990376>",
+        "Chance": 0.2
+      },
+      {
+        "Emote": "<:kekPoint:963854834795184158>",
+        "Chance": 0.2
+      },
+      {
+        "Emote": "<a:partyParrot:792709626726318091>",
+        "Chance": 0.8
+      }
+    ],
+    "RandomEmotes": [
+      {
+        "Emote": "<:pepeComfy:966290245668700191>",
+        "Chance": 0.005
+      },
+      {
+        "Emote": "\uD83D\uDE14",
+        "Chance": 0.001
+      },
+      {
+        "Emote": "<:FeelsCoffeeMan:557618333990780959>",
+        "Chance": 0.005
+      },
+      {
+        "Emote": "<:hype:527072821135015936>",
+        "Chance": 0.001
+      },
+      {
+        "Emote": "<:iAmAgony:880963270318649344>",
+        "Chance": 0.001
+      },
+      {
+        "Emote": "<:despair:881299750631145492>",
+        "Chance": 0.002
+      }
+    ],
+    "WelcomeEmotes": [
+      {
+        "Emote": "<:wave:565223468962938902>",
+        "Chance": 0.3
+      }
+    ],
     "WelcomeTriggers": [
       "welcome",
       "hello",
